+++ docs/CHANGES.rst
@@ -16,6 +16,8 @@ This library does not support Python 2.7 anymore. Attempting to install
   - ``multimode`` (Python >= 3.8)
   - ``quantiles`` (Python >= 3.8)
 
+- Add a new recipe from the ``itertools`` documentation: ``nth_combination``.
+
 Version 0.8.0 (2019-10-27)
 ^^^^^^^^^^^^^^^^^^^^^^^^^^